refactor: Move \`css\` setup, vars and palettes to folder
