+++ src/Playroom/Box/Box.tsx
@@ -1,7 +1,7 @@
 import clsx, { type ClassValue } from 'clsx';
 import { forwardRef, type AllHTMLAttributes, type ElementType } from 'react';
 
-import { sprinkles, type Sprinkles } from '../sprinkles.css';
+import { sprinkles, type Sprinkles } from '../../css/sprinkles.css';
 
 interface BoxProps
   extends Omit<
