+++ src/Playroom/CodeEditor/CodeEditor.css.ts
@@ -1,8 +1,8 @@
 import { style, globalStyle, keyframes, createVar } from '@vanilla-extract/css';
 
+import { colorPaletteVars, sprinkles } from '../../css/sprinkles.css';
+import { vars } from '../../css/vars.css';
 import { toolbarItemSize } from '../ToolbarItem/ToolbarItem.css';
-import { colorPaletteVars, sprinkles } from '../sprinkles.css';
-import { vars } from '../vars.css';
 
 const minimumLineNumberWidth = '50px';
