+++ src/Playroom/Heading/Heading.css.ts
@@ -1,7 +1,7 @@
 import { style } from '@vanilla-extract/css';
 
-import { sprinkles, colorPaletteVars } from '../sprinkles.css';
-import { vars } from '../vars.css';
+import { sprinkles, colorPaletteVars } from '../../css/sprinkles.css';
+import { vars } from '../../css/vars.css';
 
 export const base = style([
   sprinkles({
