+++ src/Playroom/Text/Text.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
-import { sprinkles, colorPaletteVars } from '../sprinkles.css';
+import { sprinkles, colorPaletteVars } from '../../css/sprinkles.css';
 
 export const base = sprinkles({
   display: 'block',
