+++ src/Playroom/icons/ChevronIcon.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
-import { vars } from '../vars.css';
+import { vars } from '../../css/vars.css';
 
 export const root = style({
   transition: vars.transition.medium,
